+++ AdventOfCode2024/Helpers/Challenge.cs
@@ -2,8 +2,8 @@
 
 public abstract class Challenge : IChallenge
 {
-    public abstract int Part1();
-    public abstract int Part2();
+    public abstract long Part1();
+    public abstract long Part2();
 
     protected readonly List<string> Data = [];
